backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-openai
langchain-anthropic
pydantic
python-dotenv
mcp
langchain-mcp-adapters
